support 'loopback' as a device type.
As reported in bug 1671927, sysconfig had an issue with rendering
a loopback device.  The problem was that some as yet unknown issue was
causing the openstack config drive to parse the provided ENI file rather
than reading the network_data.json.  Parsing an ENI file would add a
a 'lo' device of type 'physical', and sysconfig was failing to render
that.

The change here is:
 a.) add a 'loopback' type rather than 'physical' for network config.
     {'name': 'lo', 'type': 'loopback', 'subnets': ['type': 'loopback']}
 b.) support skipping that type in the eni and sysconfig renderers.
 c.) make network_state just piggy back on 'physical' renderer for
     loopback (this was what was happening before).

Tests are added for eni and sysconfig renderer.

(cherry picked from commit 1a2ca7530518d819cbab7287b12f942743427e38)

Related: rhbz#1492726
Signed-off-by: Ryan McCabe <rmccabe@redhat.com>