Bounce network interface for Azure when using the built-in path.
When deploying on Azure and using only cloud-init, you must "bounce" the
network interface to trigger a DDNS update. This allows dhclient to
register the hostname with Azure so that DNS works correctly on their
private networks (i.e. between vm and vm).

The agent path was already doing the bounce so this creates parity
between the built-in path and the agent.

LP: #1674685
Resolves: rhbz#1434109
(cherry picked from commit 86715c88aab8561e1ddadac95671f6095d16f9e7)